n_fermion_sector_view: New structure `n_fermion_sector_view`
`n_fermion_sector_view` is a view of a state vector that partially models
the `StateVector` concept. This view allows to work within a subspace
with a fixed total number of fermions. Its purpose is to sacrifice some
speed and generality of `mapped_basis_view` in order to minimize the memory
footprint when working with large sectors.

Besides the class, there is a new utility function
`n_fermion_sector_basis_states()` (returns a list of all basis state
indices in a sector) and a couple of factory functions for the views,
`make_nfs_view()` and `make_const_nfs_view()`.

The `n_fermion_sector_view` unit test has been extended to cover all added
functionality.