basis_mapper: Remove methods `make_(const_)view_no_ref()`.
`make_(const_)view()` are now taking a universal reference to the
underlying state vector. If the reference is not an lvalue reference, a
`Ref = false` view will be returned. This case was previously covered by
the `*_no_ref()` functions.

The `mapped_basis_view` unit test has been extended to check the return
types of `make_(const_)view()`.